+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 - Fixed the data saving format in Aerospike
 - Fixed the rollOverAndUpdate
+- Identified a weird client bug in Aerospike. If there are no records OR only one record Predicate.or/and don't work. They need special handling. And there's barely any documentation on the same 
 
 ## [0.0.1-RC12]